requirements.txt
openai>=1.37.0
python-pptx>=0.6.23
tqdm>=4.66.0
tenacity>=8.2.3
python-dotenv